Started the frontend development.
logo, background and main salesman btn's in place.
NO FUNTIONALITY added.
had issues connecting to the stylesheet so for the time being the styles are within the html file.

not tested with localhost because of reacuring FUCKING issue with the packages or sm shit . btw i hate intelliJ. Hey we support css! But only if you have ultimate. fuck you. fall dead n die.

dont know if we use the thymeleaf to connect as well or what happened, but had issues with images being in the resource package so had to add them within the templates/home where the main html file was. sry, we fix later