Add Price changing depending on Season, Close Contract Functionality
Add Price changes depending on the season in which the contract has been started ( +60% Peak Season, +30% Mid Season, base price Low Season ).
When a Contract is created, only the available RVs will show in the Drop-down List + the selected RV will be set to isRented = true.
The only editable fields of a Contract are now : Status, KmDriven.
When a Contract's Status is set to 'Closed' it will update the fuelStatus of the RV with a random generated float and will add the kmDriven specified in the Contract to the RV's kmDriven and will calculate & update the Contract's Final Price.