Style dataforms for create/edit customer and RV
- styled createNewCustomer.html
- styled editCustomer.html
- styled createNewRV.html
- styled editRV.html

I am continuing the work on employee, but have not finished.

might have minor issues, but it didnt seem so.
I also edited the auto-fill color that appears in the input fields.

create a new css file for the specific function of creating the form box / background and positioning of content.
There are 2 main class'es to be used when creating more dataforms:
- createForm
  * for containing everything in the form and displaying it correctly
  * header, textfields, buttons
- innerCreateForm
  * for the textfields alone to align and center them